PEDSnet model data type conversions
Convert Date and Time types to DateTime (without timezone).
Convert refills, quantity, and days supply on drug_exposure and year,
month, and day of birth on person and condition occurrence count on
condition_era to numeric from integer (to limit the range of
acceptable values). Convert concept concept_level to integer to match
OMOP specification.

Fixes #73
Fixes #78
Fixes #89
Fixes #90
Fixes #92
Fixes #93

+++ PEDSnet/sqlalchemy_vocab_tables.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, Integer, Date, String, Numeric, Time
+from sqlalchemy import Column, Integer, Date, String, Numeric
 from sqlalchemy.schema import PrimaryKeyConstraint, ForeignKeyConstraint, Index, CheckConstraint, UniqueConstraint
 from sqlalchemy.ext.declarative import declarative_base
 
@@ -14,7 +14,7 @@ class Concept(Base):
 
     concept_id = Column('concept_id', Integer(), nullable=False, autoincrement=False)
     concept_name = Column('concept_name', String(256), nullable=False)
-    concept_level = Column('concept_level', Numeric(precision=38), nullable=False)
+    concept_level = Column('concept_level', Integer(), nullable=False)
     concept_class = Column('concept_class', String(60), nullable=False)
     vocabulary_id = Column('vocabulary_id', Integer(), nullable=False)
     concept_code = Column('concept_code', String(40), nullable=False)
